backend/ai_processor.py
from transformers import pipeline
import numpy as np
from typing import Dict, List, Optional
import logging
import requests
from cachetools import TTLCache
import nats
from asyncio import run as asyncio_run
import os
import datasets
from huggingface_hub import login  # Requires API key for private datasets, handle securely
import json

class AIProcessor:
    def __init__(self):
        self.emotion_analyzer = None
        self.nlp_pipeline = None
        self.logger = logging.getLogger(__name__)
        self.initialize_models()
        # Configure cache with TTL (e.g., 3600 seconds)
        self.cache = TTLCache(maxsize=100, ttl=3600)
        # Initialize NATS client connection flag
        self.nc_connected = False

    def initialize_models(self):
        """Initialize AI models"""
        try:
            # Initialize emotion analysis pipeline
            self.emotion_analyzer = pipeline(
                "text-classification",
                model="j-hartmann/emotion-english-distilroberta-base",
                return_all_scores=True
            )

            # Initialize NLP pipeline for intent classification
            self.nlp_pipeline = pipeline(
                "text-classification",
                model="facebook/bart-large-mnli"
            )

        except Exception as e:
            self.logger.error(f"Error initializing AI models: {str(e)}")
            raise

    async def connect_nats(self):
        if not hasattr(self, 'nc') or self.nc is None:
            self.nc = await nats.connect("nats://localhost:4222")
        return self.nc

    async def analyze_emotion(self, text: str) -> Dict:
        """
        Analyze emotion in text
        Returns emotion classification with confidence scores
        """
        try:
            if not self.emotion_analyzer:
                raise ValueError("Emotion analyzer not initialized")

            results = self.emotion_analyzer(text)
            # Get the emotion with highest confidence
            emotions = results[0]
            max_emotion = max(emotions, key=lambda x: x['score'])

            confidence = float(max_emotion['score'])
            result = {
                "emotion": max_emotion['label'],
                "confidence": confidence,
                "all_emotions": emotions
            }

            return self.self_reflect(confidence, json.dumps(result))

        except Exception as e:
            self.logger.error(f"Error in emotion analysis: {str(e)}")
            raise

    async def process_text(self, text: str) -> Dict:
        """
        Process text for intent classification and entity extraction
        """
        try:
            if not self.nlp_pipeline:
                raise ValueError("NLP pipeline not initialized")

            # Classify intent
            intent_result = self.nlp_pipeline(
                text,
                candidate_labels=["question", "statement", "command", "request"]
            )

            confidence = float(intent_result['scores'][0])
            result = {
                "intent": intent_result['labels'][0],
                "confidence": confidence,
                "text": text
            }

            return self.self_reflect(confidence, json.dumps(result))

        except Exception as e:
            self.logger.error(f"Error in text processing: {str(e)}")
            raise

    async def generate_response(self, 
                              user_input: str, 
                              context: Optional[List[Dict]] = None) -> Dict:
        """
        Generate response based on user input and context
        """
        try:
            # Analyze emotion in user input
            emotion_result = await self.analyze_emotion(user_input)
            
            # Process intent
            intent_result = await self.process_text(user_input)

            # Combine results for response generation
            response_data = {
                "emotion_analysis": json.loads(emotion_result),
                "intent_analysis": json.loads(intent_result),
                "generated_response": {
                    "text": "I understand your message.",  # Placeholder
                    "confidence": 0.85
                }
            }

            return response_data

        except Exception as e:
            self.logger.error(f"Error in response generation: {str(e)}")
            raise

    def update_learning(self, interaction_data: Dict) -> bool:
        """
        Update the learning model based on interaction data
        """
        try:
            # Placeholder for model updating logic
            self.logger.info("Updating model with new interaction data")
            return True

        except Exception as e:
            self.logger.error(f"Error updating learning model: {str(e)}")
            return False

    async def analyze_image_emotion(self, image_data: bytes) -> Dict:
        """
        Analyze emotion from image data (facial expression)
        """
        try:
            # Placeholder for image-based emotion analysis
            return {
                "emotion": "neutral",
                "confidence": 0.75,
                "facial_features": {
                    "eyes": "open",
                    "mouth": "neutral",
                    "eyebrows": "neutral"
                }
            }

        except Exception as e:
            self.logger.error(f"Error in image emotion analysis: {str(e)}")
            raise

    async def get_model_stats(self) -> Dict:
        """
        Get statistics about the AI model's performance
        """
        return {
            "total_processed": 1000,
            "accuracy": 0.89,
            "last_updated": "2025-04-27T05:00:00Z",
            "model_version": "1.0.0"
        }

    async def load_conceptnet_data(self, concept: str):
        nc = await self.connect_nats()
        if concept in self.cache:
            return self.cache[concept]
        url = f"http://api.conceptnet.io/c/en/{concept}"
        response = requests.get(url)
        if response.status_code == 200:
            data = response.json()
            await nc.publish("graph.update.conceptnet", json.dumps(data).encode())
            self.cache[concept] = data
            return data
        return {"error": "Failed to load data"}

    async def load_wordnet_data(self, word: str):
        nc = await self.connect_nats()
        import nltk
        from nltk.corpus import wordnet as wn
        nltk.download('wordnet', quiet=True)
        synsets = wn.synsets(word)
        await nc.publish("graph.update.wordnet", json.dumps([syn.name() for syn in synsets]).encode())
        return [syn.name() for syn in synsets]

    async def load_wikidata_data(self, entity: str):
        nc = await self.connect_nats()
        if entity in self.cache:
            return self.cache[entity]
        url = f"https://www.wikidata.org/w/api.php?action=wbgetentities&ids={entity}&format=json"
        response = requests.get(url)
        if response.status_code == 200:
            data = response.json()
            await nc.publish("graph.update.wikidata", json.dumps(data).encode())
            self.cache[entity] = data
            return data
        return {"error": "Failed to load data"}

    async def load_emotional_dataset(self, dataset_name: str):
        nc = await self.connect_nats()
        hf_token = os.getenv('HF_TOKEN')
        if hf_token:
            login(token=hf_token)
        dataset = datasets.load_dataset(dataset_name, split='train')
        await nc.publish("graph.update.emotions", json.dumps(dataset).encode())
        return dataset

    async def integrate_emotions(self):
        await self.load_emotional_dataset('empathetic_dialogues')
        await self.load_emotional_dataset('go_emotions')

    async def self_reflect(self, confidence: float, response: str) -> str:
        if confidence < 0.7:
            return f"{response} [SASOK_DOUBT: Confidence below 70%]"
        return response

    async def cleanup(self):
        if hasattr(self, 'nc') and self.nc is not None:
            await self.nc.close()

# Removed the following usage example to avoid lint errors in non-async context:
# processor = AIProcessor()
# try:
#     await processor.load_conceptnet_data("concept")
# finally:
#     await processor.cleanup()
