backend/requirements.txt
fastapi==0.68.1
uvicorn==0.22.0  # Updated from 0.15.0 for better compatibility with websockets
python-multipart==0.0.9  # Updated from 0.0.5 to a non-yanked version
web3==6.0.0  # Updated from 5.24.0 for better dependency compatibility with websockets
numpy==1.26.4  # Updated from 1.21.2 for Python 3.12 compatibility
pandas==2.2.2  # Updated from 1.3.3
scikit-learn==1.5.0  # Updated from 0.24.2
transformers==4.44.0  # Updated from 4.11.3
python-jose==3.3.0
passlib==1.7.4
python-dotenv==0.19.0
motor==2.5.1
aiohttp==3.9.5  # Updated from 3.8.1 for better compatibility and binary wheels
neo4j==5.24.0  # Updated from 5.0.0 for Python 3.11 compatibility
spacy==3.7.2
# en-core-web-sm==3.7.0  # Commented out; install via 'python -m spacy download en_core_web_sm'
requests==2.31.0  # Added for HTTP requests in emotion detector
websockets==11.0  # Updated from 9.1 for availability and to meet dependency requirements
datasets==2.14.4
huggingface-hub==0.23.2  # Updated from 0.20.3 to meet dependency requirements for transformers
nats-py==2.2.0  # Added for NATS event bus to handle inter-component communication as per user rules
